refactor(eidos): extract knowledge::fact submodule (#2642)
Move fact domain types (Fact, FactTemporal, FactProvenance, FactLifecycle,
FactAccess, EpistemicTier, KnowledgeStage, ForgetReason, FactType,
VerificationSource, VerificationStatus, VerificationRecord, StageTransition,
FactDiff) plus timestamp helpers and constants to knowledge/fact.rs.
All items re-exported from mod.rs to preserve public API.
